Move get_tempfile_authentication_directory to package-wide utils.
The jsonapi / instance part of ftw.upgrade should not import from the command part.
Therefore get_tempfile_authentication_directory is moved out of the command package.

This helps avoid things such as ``curses`` import errors when the terminal is set up.
